fix: Fixed a binding issue and not using the proper run function
To expand on this issue. It seems some weird commit jenga was happening where some people commited to my branch and then merged it in or something? Lots of weird sheninigans of the sort. I deleted most of that intrusion and fixed up an issue where irssi using a socket made it unusuable by the server if I attempted a reset. Simple fix. I explained it in more detail to my project colleagues